+++ qml/palsar_fnf.qml
@@ -1,0 +1,53 @@
+<!DOCTYPE qgis PUBLIC 'http://mrcc.com/qgis.dtd' 'SYSTEM'>
+<qgis hasScaleBasedVisibilityFlag="0" maxScale="0" version="3.12.3-București" styleCategories="AllStyleCategories" minScale="1e+08">
+  <flags>
+    <Identifiable>1</Identifiable>
+    <Removable>1</Removable>
+    <Searchable>1</Searchable>
+  </flags>
+  <customproperties>
+    <property value="false" key="WMSBackgroundLayer"/>
+    <property value="false" key="WMSPublishDataSourceUrl"/>
+    <property value="true" key="ee-image"/>
+    <property key="ee-image-bands">
+      <value>fnf</value>
+    </property>
+    <property value="2017-1-1" key="ee-image-date"/>
+    <property value="JAXA/ALOS/PALSAR/YEARLY/FNF/2017" key="ee-image-id"/>
+    <property value="25" key="ee-image-scale"/>
+    <property value="POLYGON ((-180 -90,180 -90,180 90,-180 90,-180 -90))" key="ee-image-wkt"/>
+    <property value="0" key="embeddedWidgets/count"/>
+    <property value="Value" key="identify/format"/>
+  </customproperties>
+  <pipe>
+    <rasterrenderer band="1" nodataColor="" opacity="1" alphaBand="2" classificationMin="1" type="singlebandpseudocolor" classificationMax="3">
+      <rasterTransparency/>
+      <minMaxOrigin>
+        <limits>None</limits>
+        <extent>WholeRaster</extent>
+        <statAccuracy>Estimated</statAccuracy>
+        <cumulativeCutLower>0.02</cumulativeCutLower>
+        <cumulativeCutUpper>0.98</cumulativeCutUpper>
+        <stdDevFactor>2</stdDevFactor>
+      </minMaxOrigin>
+      <rastershader>
+        <colorrampshader clip="0" classificationMode="1" colorRampType="EXACT">
+          <colorramp name="[source]" type="gradient">
+            <prop k="color1" v="215,25,28,255"/>
+            <prop k="color2" v="43,131,186,255"/>
+            <prop k="discrete" v="0"/>
+            <prop k="rampType" v="gradient"/>
+            <prop k="stops" v="0.25;253,174,97,255:0.5;255,255,191,255:0.75;171,221,164,255"/>
+          </colorramp>
+          <item value="1" alpha="255" color="#006400" label="Forest"/>
+          <item value="2" alpha="255" color="#feff99" label="Non-Forest"/>
+          <item value="3" alpha="255" color="#0000ff" label="Water"/>
+        </colorrampshader>
+      </rastershader>
+    </rasterrenderer>
+    <brightnesscontrast contrast="0" brightness="0"/>
+    <huesaturation colorizeBlue="128" grayscaleMode="0" saturation="0" colorizeStrength="100" colorizeOn="0" colorizeRed="255" colorizeGreen="128"/>
+    <rasterresampler maxOversampling="2"/>
+  </pipe>
+  <blendMode>0</blendMode>
+</qgis>